fix(contact-center): review fixes for EP-DN hydrate, consult guards, and resume
